refactor: trim dependencies
Trim dependencies to minimize the build times of user projects.
