+++ common/Cargo.toml
@@ -12,10 +12,10 @@ anyhow = { workspace = true, optional = true }
 async-trait = { workspace = true , optional = true }
 axum = { workspace = true, optional = true }
 bytes = { version = "1.3.0", optional = true }
-chrono = { workspace = true, features = ["serde"] }
+chrono = { workspace = true }
 comfy-table = { version = "6.1.3", optional = true }
 crossterm =  { version = "0.25.0", optional = true }
-headers = { workspace = true }
+headers = { workspace = true, optional = true }
 http = { workspace = true, optional = true }
 http-body = { version = "0.4.5", optional = true }
 http-serde = { version = "1.1.2", optional = true }
@@ -25,7 +25,7 @@ once_cell = { workspace = true, optional = true }
 opentelemetry = { workspace = true, optional = true }
 opentelemetry-http = { workspace = true, optional = true }
 opentelemetry-otlp = { version = "0.11.0", optional = true }
-pin-project = { workspace = true }
+pin-project = { workspace = true, optional = true }
 prost-types = { workspace = true, optional = true }
 reqwest = { version = "0.11.13", optional = true }
 rmp-serde = { version = "1.1.1", optional = true }
@@ -44,12 +44,14 @@ ttl_cache = { workspace = true, optional = true }
 uuid = { workspace = true, features = ["v4", "serde"], optional = true }
 
 [features]
-backend = ["async-trait", "axum/matched-path", "bytes", "http", "http-body", "hyper/client", "jsonwebtoken", "opentelemetry", "opentelemetry-http", "opentelemetry-otlp", "thiserror", "tower", "tower-http", "tracing-opentelemetry", "tracing-subscriber/env-filter", "tracing-subscriber/fmt", "ttl_cache"]
-display = ["comfy-table", "crossterm"]
-models = ["anyhow", "async-trait", "display", "http", "prost-types", "reqwest", "serde_json", "service", "thiserror"]
+backend = ["async-trait", "axum/matched-path", "claims", "hyper/client", "opentelemetry-otlp", "thiserror", "tower-http", "tracing-subscriber/env-filter", "tracing-subscriber/fmt", "ttl_cache"]
+claims = ["bytes", "chrono/clock", "headers", "http", "http-body", "jsonwebtoken", "opentelemetry", "opentelemetry-http", "pin-project", "tower", "tracing", "tracing-opentelemetry"]
+display = ["chrono/clock", "comfy-table", "crossterm"]
+error = ["prost-types", "serde_json", "thiserror", "uuid"]
+models = ["anyhow", "async-trait", "display", "http", "reqwest", "serde_json", "service"]
 service = ["chrono/serde", "once_cell", "rustrict", "serde/derive", "strum", "uuid"]
 tracing = ["serde_json"]
-wasm = ["http-serde", "http", "rmp-serde", "tracing", "tracing-subscriber"]
+wasm = ["chrono/clock", "http-serde", "http", "rmp-serde", "tracing", "tracing-subscriber"]
 
 [dev-dependencies]
 axum = { workspace = true }
