+++ common/src/lib.rs
@@ -1,5 +1,7 @@
 #[cfg(feature = "backend")]
 pub mod backends;
+#[cfg(feature = "claims")]
+pub mod claims;
 #[cfg(feature = "service")]
 pub mod database;
 #[cfg(feature = "service")]
@@ -37,6 +39,18 @@ pub type ApiUrl = String;
 pub type Host = String;
 #[cfg(feature = "service")]
 pub type DeploymentId = Uuid;
+
+#[cfg(feature = "error")]
+/// Errors that can occur when changing types. Especially from prost
+#[derive(thiserror::Error, Debug)]
+pub enum ParseError {
+    #[error("failed to parse UUID: {0}")]
+    Uuid(#[from] uuid::Error),
+    #[error("failed to parse timestamp: {0}")]
+    Timestamp(#[from] prost_types::TimestampError),
+    #[error("failed to parse serde: {0}")]
+    Serde(#[from] serde_json::Error),
+}
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct DatabaseReadyInfo {
