+++ gateway/src/lib.rs
@@ -344,7 +344,8 @@ pub mod tests {
     use jsonwebtoken::EncodingKey;
     use rand::distributions::{Alphanumeric, DistString, Distribution, Uniform};
     use ring::signature::{self, Ed25519KeyPair, KeyPair};
-    use shuttle_common::backends::auth::{Claim, ConvertResponse, Scope};
+    use shuttle_common::backends::auth::ConvertResponse;
+    use shuttle_common::claims::{Claim, Scope};
     use shuttle_common::models::project;
     use sqlx::SqlitePool;
     use tokio::sync::mpsc::channel;
