+++ runtime/src/provisioner_factory.rs
@@ -2,7 +2,7 @@ use std::{collections::BTreeMap, path::PathBuf, sync::Arc};
 
 use async_trait::async_trait;
 use shuttle_common::{
-    backends::{auth::ClaimService, tracing::InjectPropagation},
+    claims::{ClaimService, InjectPropagation},
     database,
     storage_manager::StorageManager,
     DatabaseReadyInfo,
